pyproject.toml
[project]
name = "warehouse-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Katherine Patricio", email = "keitpatricio@yahoo.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.0.1",
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.13",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
warehouse-optimizer = "warehouse_optimizer:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
